server: Split modules from app, refactor purchase/payments pgp dependencies
